simplify the fuzzer test
The fuzzer test is transformed into a jest test that checks that we
can start the fuzzer including the native addon. We included patches
so that libfuzzer returns when fuzzing is finished. libFuzzer calls
exit(0) by default which makes it difficult to use it in unit tests.
This would also be later useful for test framework integration.