Fix handling of Node errors in C++ code
We were using Napi::Error::ThrowAsJavaScriptException, which is intended for
situations where C++ exceptions are disabled. The critical problem with it is
that it doesn't terminate the C++ code, expecting the caller instead to return
to the event loop immediately. Our code didn't do so, i.e., we would pretend to
operate normally. This commit changes all error-handling locations to throw the
Napi::Error as a C++ exception, causing an immediate return to the event loop.