Detect timeouts in async fuzzing mode
Timeouts can be directly configured by the libFuzzer flag -timeout. In
case of async fuzzing, this was not handled correctly.

When receiving a sigalrm signal, indicating a timeout, libFuzzer checks
if the signal is handled on it's thread and not another one of the
process. In case of async fuzzing libFuzzer is started in a dedicated
thread, but the signal is received by the main one, so the check fails
and no timeout finding is produced.

In this commit we patch out the mentioned check. It's already
deactivated for some platforms and doesn't seem to cause issues.