Fix prebuild upload on windows
Prebuild upload needs a GitHub access token to upload artifacts to a
GitHub release. This token was provided via an environment variable,
which broke on windows due to a different env var access format string.

To only use one upload script command, the token is now passed in as
additional argument after "--". This way the upload works on all
operating systems.