Ignore package-lock files for example projects
In general package-lock.json files should be committed to the
repository. In case of the example projects we don't want fixed versions
but build them with the current fuzzer and it's dependencies. This leads
to constant merge conflicts without much value. Hence, lock files of the
example projects should be ignored.