Remove libFuzzer in regression mode
It is not required to run libFuzzer in the Jest regression mode. Rather
it would execute libFuzzer multiple times, which is not officially
supported. Executing the fuzz test with the particular file contents is
sufficient to trigger unresolved errors.

Mocking the Jest test function in an async fashion is somewhat more
complicated than the previous sync version but required to allow async
file handling.