Add release canary test (#525)
This adds a test to the PR CI workflow to package up jazzer.js and run it against a project that imports in similar form to what it would get if it were importing the released version rather than using local file references. This will catch the issue we had previously where an import worked locally but not in the released version.