Fix FuzzedDataProvider crash on float length (#552)
This adds checks to all functions in `FuzzedDataProvider` that accept a
parameter that could be used as the number of bytes to pull from the
fuzzed data. If any length value is not a whole number, the function
will throw an error to prevent `dataPtr` from ending up a non-whole
number. I chose this over having it do a `Math.floor` and continuing to
make clear what's going on.

The crashes we were seeing were due to `consumeNumberInRange` returning
a non-integer value, `consumeString` would accept that value as the
length of the string to get and would set `this.dataPtr` and
`this._remainingBytes` as a direct calculation based off of the length
given. I'm not sure what `this.data.toString` does with the floating
value but it successfully returned something so this call would work.
The next call to `consumeNumberInRange` would end up trying to index
`this.data` with `this.dataPtr + <some other number>` and the end result
was `this.data[<a non-integer value>]` which JS will deal with by
returning `undefined`.