core: Provide consistent exit codes
This change uses the possibility to stop the fuzzer loop provided in the patched
libFuzzer version. Incorporating this allows the removal of the stopFuzzing
functions and general simplifications of the fuzzing process.

The core fuzzing function now returns the expected exit code and found error so
that the invoking code can handle them according to its needs, e.g. exit the
process with the given code in case of CLI, and rethrow the error in case of
Jest.