Handle real segfaults in fuzz targets (#602)
This adds handling for real segfaults in native code in our fuzz
targets. Unfortunately, the only way to do that effectively was to have
the C++ code in the fuzzer detect the error, print a message, write the
crash file, and quit rather than returning control to JS so the result
isn't as nice as a normal error but should be more informative than
before.