core: Handle SIGSEGV signal only in native code
Handling of SIGSEGV in the native addon only seems to be sufficient. In case of
a segfault print an error message, dump the current input and exit the
application. This should be a real rare case to begin with and continuing to
fuzz seems like a bad idea.

SIGINT still needs to be handled in JS, as the fuzzer should shut down properly
and generate coverage and debug information.