jest: Extend Jest timeout
The Jest integration tests start another Jest instance internally and assert
that the fuzzer was able to raise findings or print expected output. As this may
take up quite some time the timeout of the outer Jest instance is high, whereas
the inner instance only executes a fixed number of runs.