feat: add per-test dictionaries
Adds a new flag "dictionaryEntries" that can be used in Jest fuzz
tests to specify dictionaries for each fuzz test individually.
In addittion, now it is possible to set the options "timeout",
"fuzzerOptions", and "sync" on a per-fuzz test basis.

This commit also refactors the way Options are updated: now every
option knows its source, and merging happens strictly based on the
priorities. The advantage is that the options can be merged in
any order, and the developer does not need to care about special
cases: everything is handled by the OptionsManager.

This also has the side-effect that now when printing options in
debugging mode, the user will know where each option originates from.

Co-authored-by: Robert Czechowski <robert@code-intelligence.com>