BREAKING CHANGE fix(FDP): remove the ability to generate printable strings
This feature made fuzzing less efficient. The best practice is to
reject non-printable characters in the fuzz target, so that the fuzzer
can pick it up; or to tell libfuzzer to generate only printable
characters using the flag "-only_ascii" if possible.

Removing generation of the constant table from the constructor now makes FDP 10x faster.